requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.3
boto==2.10.0
itsdangerous==0.23
python-memcached==1.53
wsgiref==0.1.2
